requirements.txt
pyyaml~=6.0.1
pandas~=2.1.1
numpy~=1.26.1
dash~=2.14.0
dash-bootstrap-components~=1.5.0
plotly~=5.17.0
psycopg2-binary~=2.9.9
python3-discogs-client~=2.7
Pillow~=10.0.1
gunicorn~=21.2.0
python-box~=7.1.1